pyproject.toml
[tool.poetry]
name = "data aggregator"
version = "0.1.0"
description = ""
authors = ["CtrlPlusShift <thsibtima1999@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pandas = "^2.2.2"
pandera = "^0.20.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-postgres-extra = "^2.0.8"


[tool.poetry.group.dev.dependencies]
flake8-bugbear = "^24.8.19"
flake8-pyproject = "^1.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.flake8]
ignore = ["E203", "W503", "W504"]
max-line-length = 88
exlcude = [".git", "__pycache__", "build", "dist", ".eggs"]
